# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Loyalty Card
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
...
paths:
...
  /admins:
    post:
      summary: Create admin (only for SUPERUSER)
      tags:
        - Admins
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: Get all admins
      tags:
        - Admins
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
...
    CreateAdminRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phoneNumber
        - password
        - role
      properties:
        firstName:
          type: string
          example: Petro
        lastName:
          type: string
          example: Ivanenko
        email:
          type: string
          format: email
          example: "admin@codillas.com"
        phoneNumber:
          type: string
          example: "+417612312312"
        password:
          type: string
          format: password
          writeOnly: true
          example: 'Best_secure_password'
        role:
          type: string
          enum: [ SUPERUSER, ADMIN ]
          example: ADMIN
      additionalProperties: false
    #============
    Admin:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - phoneNumber
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: aefff665-1ab1-4160-b090-24321895a7d0
          readOnly: true
        firstName:
          type: string
          example: Petro
        lastName:
          type: string
          example: Ivanenko
        email:
          type: string
          format: email
          example: "admin@codillas.com"
        phoneNumber:
          type: string
          example: "+417612312312"
        role:
          type: string
          enum: [ SUPERUSER, ADMIN ]
          example: ADMIN
        createdAt:
          type: string
          format: date-time
          example: "2026-07-13T18:32:00Z"
          description: "Дата та час створення запису"
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-13T19:32:00Z"
          description: "Дата та час оновлення запису"
          readOnly: true
    #============
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message